backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.129.0
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.1
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
